% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{addAnnotationsFromTableToSE}
\alias{addAnnotationsFromTableToSE}
\title{Add annotation to the SummerizedExperiment.}
\usage{
addAnnotationsFromTableToSE(se, data_annotation)
}
\arguments{
\item{se}{SummerizedExperiment, With counts}

\item{data_annotation}{Dataframe, Containing all annotation data}
}
\value{
se, (SummerizedExperiment) With samples, counts and annotation
}
\description{
Add annotation to the SummerizedExperiment.
}
