% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{addSamplesFromTableToSE}
\alias{addSamplesFromTableToSE}
\title{Add samples to the SummerizedExperiment.}
\usage{
addSamplesFromTableToSE(se, data_samples)
}
\arguments{
\item{se}{SummerizedExperiment, With counts}

\item{data_samples}{Dataframe, Containing all sample data}
}
\value{
se, (SummerizedExperiment) With samples and counts
}
\description{
Add samples to the SummerizedExperiment.
}
