% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{alignment_summary}
\alias{alignment_summary}
\title{Prepare data for alignment plot}
\usage{
alignment_summary(se, percent = "percent")
}
\arguments{
\item{se}{Summarized Experiment, SE to use as input with samplesheet & counts}

\item{percent}{String, Should output value be in percentages or not}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for alignment plot
}
