% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{alignmentSummary}
\alias{alignmentSummary}
\title{Creates dataframe with counts per mapping feature (aligned, not aligned, etc.).}
\usage{
alignmentSummary(se)
}
\arguments{
\item{se}{SummerizedExperiment, With samples, counts (and annotation)}
}
\value{
out, (Dataframe) With total counts per available mapping feature
}
\description{
Creates dataframe with counts per mapping feature (aligned, not aligned, etc.).
}
