% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{alignmentSummaryPlot}
\alias{alignmentSummaryPlot}
\title{Calculates Log values in the function alignmentSummary().
If percentages == TRUE percentages are calculated.
A stacked bar plot is created based on sample read counts.}
\usage{
alignmentSummaryPlot(se, sort_value = "None", perc = T)
}
\arguments{
\item{se}{SummerizedExperiment object, containing samples and counts}

\item{sort_value}{String, Sort samples based on a group}

\item{perc}{Boolean, Data in percentages}
}
\value{
p, (Plotly object) plot
}
\description{
Calculates Log values in the function alignmentSummary().
If percentages == TRUE percentages are calculated.
A stacked bar plot is created based on sample read counts.
}
