% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_plot}
\alias{bar_plot}
\title{Create bar plot with ggplot2}
\usage{
bar_plot(
  df,
  group,
  x,
  y,
  fill = NULL,
  title = "",
  xlab = "",
  ylab = "",
  colorbar = NA,
  facet = "none"
)
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{x}{String, Value to plot on X axis}

\item{y}{String, Value to plot on Y axis}

\item{fill}{String, value to fill bars by color}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}

\item{colorbar}{String, Should default colors be used}

\item{facet}{String, Should data be ordered by group}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create bar plot with ggplot2
}
