% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{barcode}
\alias{barcode}
\title{Prepare data for barcode plot}
\usage{
barcode(deTab, dge, amount, select)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}

\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}

\item{amount}{Numeric, Number of genes that should be present in plot}

\item{select}{Vector, Manual selected genes that should be added to plot}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for barcode plot
}
