% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{barcode_plot}
\alias{barcode_plot}
\title{Create barcode plot with ggplot2}
\usage{
barcode_plot(df, x, y, group, title = "", xlab = "", ylab = "")
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{x}{String, Value to plot on X axis}

\item{y}{String, Value to plot on Y axis}

\item{group}{String, Value to color/group data}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create barcode plot with ggplot2
}
