% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{barcodePlot}
\alias{barcodePlot}
\title{The DE table is sorted on FDR/adjPvalue.
The first x genes are kept.
Normalized values are extracted based on the DE genes still present.
Normalized values are stacked.
Scatter plot is created with LogCPM values per gene per sample.}
\usage{
barcodePlot(deTab, dge, color, amount, selected)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}

\item{dge}{DGE list object, containing samples and counts}

\item{color}{String, Column on wich colors should be based}

\item{amount}{Integer, The number of genes shown in plot}

\item{selected}{Vector, Extra selected rownames}
}
\value{
p, (Plotly object) plot
}
\description{
The DE table is sorted on FDR/adjPvalue.
The first x genes are kept.
Normalized values are extracted based on the DE genes still present.
Normalized values are stacked.
Scatter plot is created with LogCPM values per gene per sample.
}
