% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{bias_gc}
\alias{bias_gc}
\title{Prepare data for GC bias plot}
\usage{
bias_gc(deTab, select)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}

\item{select}{String, GC Bias to be selected from annotation, e.g., gene, transcript or exon}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for GC bias plot
}
