% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{biasPlot}
\alias{biasPlot}
\title{The confidence prediction calculated in the function gamConfidenceFit().
Confidence is calculated based on the GC or length.
Scatterplot is created with avgLog2FC and corresponding bias value.}
\usage{
biasPlot(deTab, biasColumn, log, tick, sourceId)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}

\item{biasColumn}{String, Column indicating bias values (GC or length)}

\item{log}{Boolean, Show plot in Log scale}
}
\value{
p, (Plotly object) plot
}
\description{
The confidence prediction calculated in the function gamConfidenceFit().
Confidence is calculated based on the GC or length.
Scatterplot is created with avgLog2FC and corresponding bias value.
}
