% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{cnet_data}
\alias{cnet_data}
\title{Prepare data for concept network plot}
\usage{
cnet_data(deTab, graphData, terms)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}

\item{graphData}{Dataframe, Data object with graph location}

\item{terms}{Numeric, Number of terms that should be present in plot}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for concept network plot
}
