% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{cnetPlotly}
\alias{cnetPlotly}
\title{Genes are gathered from enrichment result from specific pathway.
The geneList is prepared with the right LogFC values and links of genes between multiple pathways.
The number of pathways that are shown is defined with cnet_slider.
Igraph object is made to create links of genes between multiple pathways.
Colors of dots are based on LogFC.}
\usage{
cnetPlotly(enrich, geneSets, deTab)
}
\arguments{
\item{enrich}{Enrich result, A enrichment results}

\item{geneSets}{Martrix object, containing all pathways with corresponding genes}

\item{deTab}{Dataframe, with all analysis results}
}
\value{
g, (Igraph object) containing links of genes multiple pathways
}
\description{
Genes are gathered from enrichment result from specific pathway.
The geneList is prepared with the right LogFC values and links of genes between multiple pathways.
The number of pathways that are shown is defined with cnet_slider.
Igraph object is made to create links of genes between multiple pathways.
Colors of dots are based on LogFC.
}
