% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{complexity}
\alias{complexity}
\title{Prepare data for complexity plot}
\usage{
complexity(se, rank = 1000)
}
\arguments{
\item{se}{Summarized Experiment, SE to use as input with samplesheet & counts}

\item{rank}{Numeric, How many genes should be plotted}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for complexity plot
}
