% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{complexityData}
\alias{complexityData}
\title{Creates dataframe with number of reads per gene, ordered on genes with the most reads assigned.}
\usage{
complexityData(se, max)
}
\arguments{
\item{se}{SummerizedExperiment, With samples, counts (and annotation)}

\item{max}{Integer, The maximum rank that is used}
}
\value{
out, (Dataframe) With total read counts per gene
}
\description{
Creates dataframe with number of reads per gene, ordered on genes with the most reads assigned.
}
