% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{complexityPlot}
\alias{complexityPlot}
\title{Calculates count data per rank in the function complexityData().
If percentages == TRUE percentages are calculated.
A dot line plot is created based on the number of reads at rank.}
\usage{
complexityPlot(se, color = "None", perc = T, rank = 1000)
}
\arguments{
\item{se}{SummerizedExperiment object, containing samples and counts}

\item{color}{String, Sort samples based on a group}

\item{perc}{Boolean, Data in percentages}

\item{rank}{Integer, The number of genes/rank (min=10)}
}
\value{
p, (Plotly object) plot
}
\description{
Calculates count data per rank in the function complexityData().
If percentages == TRUE percentages are calculated.
A dot line plot is created based on the number of reads at rank.
}
