% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{count_dist}
\alias{count_dist}
\title{Prepare data for count distribution plot}
\usage{
count_dist(dge)
}
\arguments{
\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for count distribution plot
}
