% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{countDistributionBoxPlot}
\alias{countDistributionBoxPlot}
\title{Stacks count data from dge list in the function stackDge().
A box plot is created based on Log2CPM counts of the samples.}
\usage{
countDistributionBoxPlot(dge)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}
}
\value{
p, (Plotly object) plot
}
\description{
Stacks count data from dge list in the function stackDge().
A box plot is created based on Log2CPM counts of the samples.
}
