% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{createContrast}
\alias{createContrast}
\title{Create contrast by finding matrix values in the design matrix.
Values are assigned based if present in matrix (v1 or v2) else 0.}
\usage{
createContrast(design, matrix_v1, matrix_v2)
}
\arguments{
\item{design}{String, in use design matrix}

\item{matrix_v1}{Vector, containing the left matrix values}

\item{matrix_v2}{Vector, containing the right matrix values}
}
\value{
contrast, (vector) containing the contrasts (0, 1, -1)
}
\description{
Create contrast by finding matrix values in the design matrix.
Values are assigned based if present in matrix (v1 or v2) else 0.
}
