% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{createDesign}
\alias{createDesign}
\title{Create vector from two design values and two matrix values.
Check if all values are the same then design will start with ~0.
Get column names based on corresponding values in matrix.
Paste column names together with +.
Create String from vector seprated by ''.}
\usage{
createDesign(dge, dbase, dvalue, matrix_v1, matrix_v2)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{dbase}{Vector, base design column}

\item{dvalue}{Vector, other design columns}

\item{matrix_v1}{Vector, containing the left matrix values}

\item{matrix_v2}{Vector, containing the right matrix values}
}
\value{
getdesign, (String) String with the design formula
}
\description{
Create vector from two design values and two matrix values.
Check if all values are the same then design will start with ~0.
Get column names based on corresponding values in matrix.
Paste column names together with +.
Create String from vector seprated by ''.
}
