% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{createMatrix}
\alias{createMatrix}
\title{Getbase column and use this to create right matrix.
Loop through base, nested columns and matrix to create full matrix.}
\usage{
createMatrix(dge, dbase, dvalue, matrix)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{dbase}{Vector, base design column}

\item{dvalue}{Vector, other design columns}

\item{matrix}{Vector, containing the matrix values}
}
\value{
dge, (DGE list object) containing samples and counts
}
\description{
Getbase column and use this to create right matrix.
Loop through base, nested columns and matrix to create full matrix.
}
