% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{de_ratio}
\alias{de_ratio}
\title{Prepare data for expression ratio plot}
\usage{
de_ratio(deTab)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for expression ratio plot
}
