% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{dendro_data}
\alias{dendro_data}
\title{Prepare data for dendrogram plot}
\usage{
dendro_data(dge)
}
\arguments{
\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for dendrogram plot
}
