% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{deRatioPlot}
\alias{deRatioPlot}
\title{The number of unique values in column 'DE' are extracted.
Percentages of these values are calculated and renamed.
Barplot is created with DE results based on percentages.}
\usage{
deRatioPlot(deTab)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}
}
\value{
p, (Plotly object) plot
}
\description{
The number of unique values in column 'DE' are extracted.
Percentages of these values are calculated and renamed.
Barplot is created with DE results based on percentages.
}
