% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{emap_plotly}
\alias{emap_plotly}
\title{Pathways are gathered from enrichment result with pathway description.
A Igraph object is created with links between pathways.
Colors of dots are based on p-value.}
\usage{
emap_plotly(enrich)
}
\arguments{
\item{enrich}{Enrich result, A enrichment results}
}
\value{
g, (Igraph object) containing links between pathways
}
\description{
Pathways are gathered from enrichment result with pathway description.
A Igraph object is created with links between pathways.
Colors of dots are based on p-value.
}
