% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{enrich_bar}
\alias{enrich_bar}
\title{Prepare data for enriched bar plot}
\usage{
enrich_bar(enrich, amount)
}
\arguments{
\item{enrich}{Dataframe, Dataframe with enrichment results from analysis}

\item{amount}{Numeric, Number of terms that should be present in plot}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for enriched bar plot
}
