% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{enrich_barDE}
\alias{enrich_barDE}
\title{Prepare data for enriched bar plot with DE genes}
\usage{
enrich_barDE(enrich, amount, deTab)
}
\arguments{
\item{enrich}{Dataframe, Dataframe with enrichment results from analysis}

\item{amount}{Numeric, Number of terms that should be present in plot}

\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for enriched bar plot with DE genes
}
