% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichBarplot}
\alias{enrichBarplot}
\title{The enrichment results is filtered based on the number shown of pathways.
Descriptions of the pathways are added and the color scale is calculated.
A Barplot is created with a bar per pathway.}
\usage{
enrichBarplot(enrich, amount, value)
}
\arguments{
\item{enrich}{Enrich result, A enrichment results}

\item{amount}{Integer, Value with the number of pathways to show}

\item{value}{String, Value with the column name to be used (p-, q- or adjP value)}
}
\value{
p, (Plotly object) plot
}
\description{
The enrichment results is filtered based on the number shown of pathways.
Descriptions of the pathways are added and the color scale is calculated.
A Barplot is created with a bar per pathway.
}
