% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{enrichDE}
\alias{enrichDE}
\title{The enrichment results is filtered based on the number shown of pathways.
Number of up/down regulated genes is retrieved and added per term.
A Barplot is created with a bar per pathway.
For each pathway there is bar with up/down regulated genes.}
\usage{
enrichDE(enrich, deTab, amount)
}
\arguments{
\item{enrich}{Enrich result, A enrichment results}

\item{deTab}{Dataframe, with all analysis results}

\item{amount}{Integer, Value with the number of pathways to show}
}
\value{
p, (Plotly object) plot
}
\description{
The enrichment results is filtered based on the number shown of pathways.
Number of up/down regulated genes is retrieved and added per term.
A Barplot is created with a bar per pathway.
For each pathway there is bar with up/down regulated genes.
}
