% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{extract_geneSets}
\alias{extract_geneSets}
\title{Collect genes within a pathway.
Gets genes from enrichment result.
geneSets names are set based on ID and description.
duplicated pathways are removed from the set.}
\usage{
extract_geneSets(enrich, slider, selected)
}
\arguments{
\item{enrich}{Enrichment object, containing enrichment results}

\item{slider}{Integer, Updated number of pathways to show}

\item{selected}{Vector, Contains manually selected pathways}
}
\value{
geneSets, (vector) Genes found in corresponding pathways
}
\description{
Collect genes within a pathway.
Gets genes from enrichment result.
geneSets names are set based on ID and description.
duplicated pathways are removed from the set.
}
