% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{fix_violin_hover}
\alias{fix_violin_hover}
\title{Fix hover info for violin plot, as it cant find CPM values}
\usage{
fix_violin_hover(plot)
}
\arguments{
\item{plot}{List object, List with plot object from ggplotly}
}
\value{
plot, List object with fixed hover
}
\description{
Fix hover info for violin plot, as it cant find CPM values
}
