% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{gamConfidenceFit}
\alias{gamConfidenceFit}
\title{Calculates a confidence fit for various plots.
Model used is GAM to calculate predictions.}
\usage{
gamConfidenceFit(deTab, biasColumn)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}

\item{biasColumn}{String, Value on which to calculate confidence}
}
\value{
prediction, (Vector) With coordinates of prediction locations (line plot)
}
\description{
Calculates a confidence fit for various plots.
Model used is GAM to calculate predictions.
}
