% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{gene_strand}
\alias{gene_strand}
\title{Prepare data for gene strand bias plot}
\usage{
gene_strand(deTab)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for gene strand bias plot
}
