% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{geneStrandBar}
\alias{geneStrandBar}
\title{The number of genes are grouped by strand and DE.
A barplot is created with the # genes divided between + and - strand.}
\usage{
geneStrandBar(deTab)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}
}
\value{
p, (Plotly object) plot
}
\description{
The number of genes are grouped by strand and DE.
A barplot is created with the # genes divided between + and - strand.
}
