% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de.R
\name{get_dendrogram_data}
\alias{get_dendrogram_data}
\title{Get heigth information from tree object.
Get labels/names from tree ends.
Calculate start and end positions of tree branches
Save line coordinates in dataframe}
\usage{
get_dendrogram_data(
  tree,
  labels = NULL,
  horiz = FALSE,
  reverseDirection = FALSE,
  hang = 0.1,
  xlab = "",
  ylab = "",
  axes = TRUE,
  cex.labels = 1,
  ...,
  adjustRange = FALSE
)
}
\arguments{
\item{tree}{Hclust object, tree object}
}
\value{
new, (Dataframe) dataframe with all dendrogram line coordinates
}
\description{
Get heigth information from tree object.
Get labels/names from tree ends.
Calculate start and end positions of tree branches
Save line coordinates in dataframe
}
