% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{get_geneList}
\alias{get_geneList}
\title{Get LogFC from deTab.
Add entrez id to geneList.
Sort geneList based on LogFC (decreasing).
Remove NA and duplicate names.}
\usage{
get_geneList(deTab)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}
}
\value{
geneList, (Vector) Named vector containing gene names and LogFC values
}
\description{
Get LogFC from deTab.
Add entrez id to geneList.
Sort geneList based on LogFC (decreasing).
Remove NA and duplicate names.
}
