% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{get_organismID}
\alias{get_organismID}
\title{The gene IDs are gathered from DE table.
The first part of the ID name is kept.}
\usage{
get_organismID(deTab)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}
}
\value{
id, (String) Organism ID value
}
\description{
The gene IDs are gathered from DE table.
The first part of the ID name is kept.
}
