% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{getCount}
\alias{getCount}
\title{Count the read counts per feature to find number of reads that are aligned, not aligned, etc.}
\usage{
getCount(x, raw)
}
\arguments{
\item{x}{Containing the value/label on which to count reads}

\item{raw}{Dataframe, Containing count data}
}
\value{
Integer, With total counted reads
}
\description{
Count the read counts per feature to find number of reads that are aligned, not aligned, etc.
}
