% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{heat_de}
\alias{heat_de}
\title{Prepare data for heatmap plot}
\usage{
heat_de(dge, deTab, amount)
}
\arguments{
\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}

\item{deTab}{Dataframe, Dataframe with expression results from analysis}

\item{amount}{Numeric, Number of genes that should be present in plot}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for heatmap plot
}
