% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{heat_terms}
\alias{heat_terms}
\title{Prepare data for enrichment heatmap plot}
\usage{
heat_terms(geneSets, deTab)
}
\arguments{
\item{geneSets}{Dataframe, GeneSets from corresponding terms}

\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for enrichment heatmap plot
}
