% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{heatmap_plot}
\alias{heatmap_plot}
\title{Create heatmap plot with ggplot2}
\usage{
heatmap_plot(df, group, x, y, fill, title = "", xlab = "", ylab = "")
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{group}{String, Value to color/group data}

\item{x}{String, Value to plot on X axis}

\item{y}{String, Value to plot on Y axis}

\item{fill}{String, value to fill bars by color}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create heatmap plot with ggplot2
}
