% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{heatplotly}
\alias{heatplotly}
\title{The genesets are gathered.
Side inforamtion like fc is added.
Frequency of genes and pathways is calculated and sorted upon.
Heatmap is created with genes, pathways, and log2FC.}
\usage{
heatplotly(geneSets, deTab)
}
\arguments{
\item{geneSets}{Martrix object, containing all pathways with corresponding genes}

\item{deTab}{Dataframe, with all analysis results}
}
\value{
p, (Plotly object) plot
}
\description{
The genesets are gathered.
Side inforamtion like fc is added.
Frequency of genes and pathways is calculated and sorted upon.
Heatmap is created with genes, pathways, and log2FC.
}
