% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{informationBox}
\alias{informationBox}
\title{Template for plot information.}
\usage{
informationBox(infoText)
}
\arguments{
\item{infoText}{String, Explanation of a plot}
}
\value{
Shiny Box object
}
\description{
Template for plot information.
}
