% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{line_plot}
\alias{line_plot}
\title{Create line plot with ggplot2}
\usage{
line_plot(
  df,
  x,
  y,
  text = NA,
  group,
  plot = "",
  title = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{x}{String, Value to plot on X axis}

\item{y}{String, Value to plot on Y axis}

\item{text}{String, Hover info text}

\item{group}{String, Value to color/group data}

\item{plot}{String, Extra condition for complexity plot}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create line plot with ggplot2
}
