% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{list2df}
\alias{list2df}
\title{The inputList is converted to a dataframe format.
Dataframe object is created based on a vector/list.
Igraph object is created based on a vector/list.}
\usage{
list2df(inputList)
}
\arguments{
\item{inputList}{Vector, containing Gene names and LogFC}
}
\value{
ldf, (Dataframe) With all in use genes
}
\description{
The inputList is converted to a dataframe format.
Dataframe object is created based on a vector/list.
Igraph object is created based on a vector/list.
}
