% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{list2graph}
\alias{list2graph}
\title{The inputList is converted to a dataframe.
The dataframe is converted to a graph.
Igraph object is created based on a vector/list.}
\usage{
list2graph(inputList)
}
\arguments{
\item{inputList}{Vector, containing Gene names and LogFC}
}
\value{
g, (Igraph object) Graph with links between genes
}
\description{
The inputList is converted to a dataframe.
The dataframe is converted to a graph.
Igraph object is created based on a vector/list.
}
