% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{ma}
\alias{ma}
\title{Prepare data for MA plot}
\usage{
ma(deTab)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for MA plot
}
