% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{ma_plot}
\alias{ma_plot}
\title{The confidence prediction calculated in the function gamConfidenceFit().
Confidence is calculated based on avgLog2CPM.
Scatterplot is created with DE results and a line showing confidence.}
\usage{
ma_plot(deTab)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}
}
\value{
p, (Plotly object) plot
}
\description{
The confidence prediction calculated in the function gamConfidenceFit().
Confidence is calculated based on avgLog2CPM.
Scatterplot is created with DE results and a line showing confidence.
}
