% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{mds_clust}
\alias{mds_clust}
\title{Prepare data for MDS plot}
\usage{
mds_clust(dge)
}
\arguments{
\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for MDS plot
}
