% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{multidimensionalScalingPlot}
\alias{multidimensionalScalingPlot}
\title{Calculates required values with 'plotMDS' method.
The plot is created with plotly with values retrieved from the mds object.
Plot is colored based on the selected column.}
\usage{
multidimensionalScalingPlot(dge, color, sourceId)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{color}{String, Column on wich colors should be based}

\item{sourceId}{plot ID, depends on raw/normalized counts}
}
\value{
p, (Plotly object) plot
}
\description{
Calculates required values with 'plotMDS' method.
The plot is created with plotly with values retrieved from the mds object.
Plot is colored based on the selected column.
}
