% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{network_plot}
\alias{network_plot}
\title{Create network plot with ggplot2}
\usage{
network_plot(
  df,
  title = "",
  xlab = "",
  ylab = "",
  label1 = TRUE,
  label2 = FALSE
)
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}

\item{label1}{Boolean, Should terms be labeled in plot}

\item{label2}{Boolean, Should genes be labeled in plot}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create network plot with ggplot2
}
