% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{overlap_ratio}
\alias{overlap_ratio}
\title{Overlap is calculated and define between x and y.
The number of overlaps is returned.
Igraph object is created based on a vector/list.}
\usage{
overlap_ratio(x, y)
}
\arguments{
\item{x}{Vector, genes relative to x axis}

\item{y}{Vector, genes relative to y axis}
}
\value{
n, (Integer) Number of overlapping values
}
\description{
Overlap is calculated and define between x and y.
The number of overlaps is returned.
Igraph object is created based on a vector/list.
}
