% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{plot_module_trait_relation_heat}
\alias{plot_module_trait_relation_heat}
\title{Create heatmap with trait data compared to module relations.
Show module name, correlation between trait and module, and the trait p-value}
\usage{
plot_module_trait_relation_heat(moduleTraitCor, moduleTraitPvalue)
}
\arguments{
\item{moduleTraitCor}{Dataframe, dataframe with all trait correlation values}

\item{moduleTraitPvalue}{Dataframe, dataframe with all trait p-values}
}
\value{
p, (Plotly object) plot
}
\description{
Create heatmap with trait data compared to module relations.
Show module name, correlation between trait and module, and the trait p-value
}
