% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{plot_modules}
\alias{plot_modules}
\title{Barplot with module sizes.
Module sizes are defined in percentages and colors are added
Stacked barplot is returned}
\usage{
plot_modules(color)
}
\arguments{
\item{color}{Vector, all module colors of all used genes}
}
\value{
p, (Plotly object) plot
}
\description{
Barplot with module sizes.
Module sizes are defined in percentages and colors are added
Stacked barplot is returned
}
