% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{plot_power}
\alias{plot_power}
\title{Create dot plot with power values (scale independence).
A R^2 cutoff line indicates the min. power levels.}
\usage{
plot_power(soft, cutoff)
}
\arguments{
\item{soft}{Dataframe, soft threshold values per power}

\item{cutoff}{Integer, R^2 cutoff value}
}
\value{
p, (Plotly object) plot
}
\description{
Create dot plot with power values (scale independence).
A R^2 cutoff line indicates the min. power levels.
}
