% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{plot_soft}
\alias{plot_soft}
\title{Create dot plot with soft power values (mean connectivity).}
\usage{
plot_soft(soft)
}
\arguments{
\item{soft}{Dataframe, soft threshold values per power}
}
\value{
p, (Plotly object) plot
}
\description{
Create dot plot with soft power values (mean connectivity).
}
