% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{plot_trait_heat}
\alias{plot_trait_heat}
\title{Create heatmap with trait data.
Show sample, trait value (column) and the exact trait value}
\usage{
plot_trait_heat(trait)
}
\arguments{
\item{trait}{Dataframe, dataframe with all trait values}
}
\value{
p, (Plotly object) plot
}
\description{
Create heatmap with trait data.
Show sample, trait value (column) and the exact trait value
}
