% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcgna.R
\name{plotly_dendro_heat}
\alias{plotly_dendro_heat}
\title{Create heatmap from gene clusters.
Show genes vs genes colored by adjacency}
\usage{
plotly_dendro_heat(data_TOM)
}
\arguments{
\item{data_TOM}{WGCNA object, Topology Overlap Matrix}
}
\value{
p, (Plotly object) plot
}
\description{
Create heatmap from gene clusters.
Show genes vs genes colored by adjacency
}
