% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{plotly_dendrogram}
\alias{plotly_dendrogram}
\title{Get dendrogram data dendro_data().
Plot a dendrogram.}
\usage{
plotly_dendrogram(d, color, color_list)
}
\arguments{
\item{d}{Hclust object, tree object}

\item{color}{String, color samples}

\item{color_list}{Vector, Color genes}
}
\value{
p, (Plotly object) plot
}
\description{
Get dendrogram data dendro_data().
Plot a dendrogram.
}
