% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{plotlyGraph}
\alias{plotlyGraph}
\title{Graph layout is set with kamada kawai (kk).
Plots are created with dots as a gene or pathway.
Lines between dots show connection of these genes and/or pathways.
The geneList is prepared with the right LogFC values and links between genes between multiple pathways.
The number of pathways that are shown is defined with cnet_slider.
Igraph object is made to create links of genes between multiple pathways.
Colors of dots are based on LogFC or P-value.}
\usage{
plotlyGraph(g, pwName, getColor, cnet, annoP, annoG)
}
\arguments{
\item{g}{Igraph object, containing graph data from genes and/or pathays}

\item{pwName}{String, Name of the shown pathway}

\item{getColor}{String, Color given to dots (LogFC or P-Values)}

\item{cnet}{Integer, Number of cnet pathways to show}

\item{annoP}{Boolean, Show pathway labels yes or no}

\item{annoG}{Boolean, Show gene labels yes or no}
}
\value{
p, (Plotly object) plot
}
\description{
Graph layout is set with kamada kawai (kk).
Plots are created with dots as a gene or pathway.
Lines between dots show connection of these genes and/or pathways.
The geneList is prepared with the right LogFC values and links between genes between multiple pathways.
The number of pathways that are shown is defined with cnet_slider.
Igraph object is made to create links of genes between multiple pathways.
Colors of dots are based on LogFC or P-value.
}
