% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{pvalue_data}
\alias{pvalue_data}
\title{Prepare data for pvalue plot}
\usage{
pvalue_data(deTab)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for pvalue plot
}
