% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{pValuePlot}
\alias{pValuePlot}
\title{The Pvalues are rounded on two decimals.
All occurences of pvalues are counted.
Bar plot is created with the p value vs occurence.}
\usage{
pValuePlot(deTab)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}
}
\value{
p, (Plotly object) plot
}
\description{
The Pvalues are rounded on two decimals.
All occurences of pvalues are counted.
Bar plot is created with the p value vs occurence.
}
