% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{readCountsFromTable}
\alias{readCountsFromTable}
\title{Add all counts that are in samples in SummerizedExperiment.}
\usage{
readCountsFromTable(data_counts, data_samples)
}
\arguments{
\item{data_counts}{Dataframe, Containing all raw count data}

\item{data_samples}{Dataframe, Containing all sample data}
}
\value{
se, (SummerizedExperiment) With counts
}
\description{
Add all counts that are in samples in SummerizedExperiment.
}
