% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdown.R
\name{relevelSamples}
\alias{relevelSamples}
\title{Create vector from two matrix values.
Try to relevel samples on values that are not present in matrix.}
\usage{
relevelSamples(dge, dbase, dvalue, matrix_v1, matrix_v2)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{dbase}{Vector, base design column}

\item{dvalue}{Vector, other design columns}

\item{matrix_v1}{Vector, containing the left matrix values}

\item{matrix_v2}{Vector, containing the right matrix values}
}
\value{
dge, (DGE list object) containing samples and counts
}
\description{
Create vector from two matrix values.
Try to relevel samples on values that are not present in matrix.
}
