% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Create scatter plot with ggplot2}
\usage{
scatter_plot(
  df,
  size = 1.5,
  key = NA,
  index = NA,
  x,
  y,
  group,
  title = "",
  xlab = "",
  ylab = "",
  scale = NA
)
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{size}{Numeric, size of dots (default 1.5)}

\item{key}{String, Value for point selection by plotly}

\item{index}{Vector, vector of items to select from dataframe}

\item{x}{String, Value to plot on X axis}

\item{y}{String, Value to plot on Y axis}

\item{group}{String, Value to color/group data}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}

\item{scale}{String, Should X-axis be scaled in log10}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create scatter plot with ggplot2
}
