% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_functions.R
\name{stackDge}
\alias{stackDge}
\title{Stacks total DGE counts based on: mapping feature (aligned, not aligned, etc.), sample and LogCPM.}
\usage{
stackDge(dge)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{max}{Integer, The maximum rank that is used}
}
\value{
count, (Dataframe) With mapping feature, sample and LogCPM
}
\description{
Stacks total DGE counts based on: mapping feature (aligned, not aligned, etc.), sample and LogCPM.
}
