% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{tsne_data}
\alias{tsne_data}
\title{Prepare data for tSNE plot}
\usage{
tsne_data(dge)
}
\arguments{
\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for tSNE plot
}
