% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{tsnePlot}
\alias{tsnePlot}
\title{Perplexity is set (max 30 & min 1), this depends on the number of samples
tsne model is calculated.
Scatter plot is created based on selected PCs.}
\usage{
tsnePlot(dge, color)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{color}{String, Column on which colors should be based}
}
\value{
p, (Plotly object) plot
}
\description{
Perplexity is set (max 30 & min 1), this depends on the number of samples
tsne model is calculated.
Scatter plot is created based on selected PCs.
}
