% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{update_n}
\alias{update_n}
\title{Sets a new n value.
n stands for the number of pathways to show.}
\usage{
update_n(x, showCategory)
}
\arguments{
\item{x}{Enrichment object, containing enrichment results}

\item{showCategory}{Integer, number of pathways}
}
\value{
n, (Integer) Updated number of pathways to show
}
\description{
Sets a new n value.
n stands for the number of pathways to show.
}
