% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{variableHeatmapPlot}
\alias{variableHeatmapPlot}
\title{LogCPM values of counts are calculated.
Variance is calculated and the first x genes are kept.
Heatmap with the values left in high_var_cpm.}
\usage{
variableHeatmapPlot(dge, amount)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{amount}{Integer, The number of genes shown in plot}
}
\value{
p, (Plotly object) plot
}
\description{
LogCPM values of counts are calculated.
Variance is calculated and the first x genes are kept.
Heatmap with the values left in high_var_cpm.
}
