% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{variancePcaPlot}
\alias{variancePcaPlot}
\title{Columns and rows from DGE list are turned.
PCA is calculated with prcomp.
PC percentages are calculated.
Barplot is created with PC percentages.}
\usage{
variancePcaPlot(dge)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}
}
\value{
p, (Plotly object) plot
}
\description{
Columns and rows from DGE list are turned.
PCA is calculated with prcomp.
PC percentages are calculated.
Barplot is created with PC percentages.
}
