% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{violin_dist}
\alias{violin_dist}
\title{Prepare data for violin distribution plot}
\usage{
violin_dist(dge)
}
\arguments{
\item{dge}{DGE List, DGE List with samplesheet, count data & annotation}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for violin distribution plot
}
