% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prep.R
\name{volcano}
\alias{volcano}
\title{Prepare data for volcano plot}
\usage{
volcano(deTab)
}
\arguments{
\item{deTab}{Dataframe, Dataframe with expression results from analysis}
}
\value{
plot_data, Dataframe with cleaned data
}
\description{
Prepare data for volcano plot
}
