% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{volcanoPlot}
\alias{volcanoPlot}
\title{Creates scatter plot with avgLog2FC vs -log10(FDR).
Two lines are generated indicating LogFC cutoff and p value cutoff.}
\usage{
volcanoPlot(deTab, LogCut, PCut)
}
\arguments{
\item{deTab}{Dataframe, with all analysis results}

\item{LogCut}{Integer, LogFC cutoff for line}

\item{PCut}{Integer, Pvalue cutoff for line}
}
\value{
p, (Plotly object) plot
}
\description{
Creates scatter plot with avgLog2FC vs -log10(FDR).
Two lines are generated indicating LogFC cutoff and p value cutoff.
}
