% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{voomPlot}
\alias{voomPlot}
\title{Calculates required values with 'voom' method.
The plot is created with plotly with values retrieved from the voom object.}
\usage{
voomPlot(dge, sourceId)
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{sourceId}{plot ID, depends on raw/normalized counts}
}
\value{
p, (Plotly object) plot
}
\description{
Calculates required values with 'voom' method.
The plot is created with plotly with values retrieved from the voom object.
}
