% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{bar_plot}
\alias{bar_plot}
\title{Create bar plot with ggplot2}
\usage{
bar_plot(
  df,
  x,
  y,
  text = NA,
  group,
  fill = NULL,
  colorbar = NA,
  rev = FALSE,
  facet = "none",
  plot = NA,
  title = "",
  xlab = "",
  ylab = ""
)
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{x}{String, Value to plot on X axis}

\item{y}{String, Value to plot on Y axis}

\item{text}{String, Hover info text}

\item{group}{String, Value to color/group data}

\item{fill}{String, value to fill bars by color}

\item{colorbar}{String, Should default colors be used}

\item{rev}{Boolean, Plot bars in reverse order (alignment sum) for correct colors}

\item{facet}{String, Should data be ordered by group}

\item{plot}{String, Extra condition for ratio plot}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create bar plot with ggplot2
}
