% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de_plots.R
\name{countDistributionLinePlot}
\alias{countDistributionLinePlot}
\title{Stacks count data from dge list in the function stackDge().
A line plot is created based on Log2CPM and the density of the counts.
A dot line plot is created based on the number of reads at rank.}
\usage{
countDistributionLinePlot(dge, color = "None")
}
\arguments{
\item{dge}{DGE list object, containing samples and counts}

\item{color}{String, Sort samples based on a group}
}
\value{
p, (Plotly object) plot
}
\description{
Stacks count data from dge list in the function stackDge().
A line plot is created based on Log2CPM and the density of the counts.
A dot line plot is created based on the number of reads at rank.
}
