% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{violin_plot}
\alias{violin_plot}
\title{Create violin plot with ggplot2}
\usage{
violin_plot(df, group, title = "", xlab = "", ylab = "")
}
\arguments{
\item{df}{Dataframe, Dataframe to create plot with}

\item{title}{String, Title for plot}

\item{xlab}{String, Label for X axis}

\item{ylab}{String, Label for Y axis}

\item{plot}{String, Extra condition for complexity plot}
}
\value{
gg, Plot object (ggplot2)
}
\description{
Create violin plot with ggplot2
}
